+++ src/Concordium/Client/LegacyCommands.hs
@@ -117,7 +117,6 @@ data LegacyCmd
     { legacyBlockHash :: !(Maybe Text) }
   | GetCryptographicParameters
     { legacyBlockHash :: !(Maybe Text) }
-  | GetChainParameters
   deriving (Show)
 
 legacyProgramOptions :: Parser LegacyCmd
@@ -127,7 +126,6 @@ legacyProgramOptions =
      getTransactionStatusCommand <>
      getTransactionStatusInBlockCommand <>
      getConsensusInfoCommand <>
-     getChainParametersCommand <>
      getBlockInfoCommand <>
      getBlockSummaryCommand <>
      getBlocksAtHeightCommand <>
@@ -230,14 +228,6 @@ getConsensusInfoCommand =
     (info
        (pure GetConsensusInfo)
        (progDesc "Query the gRPC server for the consensus information."))
-
-getChainParametersCommand :: Mod CommandFields LegacyCmd
-getChainParametersCommand =
-    command
-      "GetChainParameters"
-      (info
-        (pure GetChainParameters)
-        (progDesc "Query the gRPC server for the blockchain parameters."))
 
 getBlockInfoCommand :: Mod CommandFields LegacyCmd
 getBlockInfoCommand =